日付文字列パーサの強化

+++ src/core/risse/src/risseDateParser.h
@@ -62,7 +62,7 @@ public:
 	//! @brief		トークンを得る
 	//! @param		value		値の格納先
 	//! @return		トークンID
-	int GetToken(int & val);
+	int GetToken(long & val);
 
 };
 //---------------------------------------------------------------------------
